Make asset rendering manual and add rendered assets
