+++ .gitattributes
@@ -1,1 +1,2 @@
 *.blend filter=lfs diff=lfs merge=lfs -text
+*.png filter=lfs diff=lfs merge=lfs -text
