+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = ["logic"]
+exclude = ["tools/render-assets"]
 
 [package]
 name = "plenty-of-fish-in-the-sea"
@@ -17,9 +18,6 @@ num-rational = {version = "0.4"}
 [build-dependencies]
 tar = "0.4"
 built = { version = "0.5", features = ["git2", "chrono"] }
-toml = "0.5"
-serde = { version = "1.0", features = ["derive"] }
-
 
 
 [profile.release]
