+++ assets/rendered/fish-04.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0120b24ac777b8db96c4c795849f61186c98957cbc3c1a62c065919970919c37
+size 8756
