+++ assets/rendered/fish-06.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f5ef014fb232590f78958020f5069da57d4bbc8e7d5d01f43beb928cbbfa23d5
+size 12083
