+++ assets/rendered/ship-00.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:aa6763a79860bd7eca94a0a5a5484318560dfad76f5772bc16bb0b89c0cb6ffa
+size 290262
