+++ assets/rendered/ship-01.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7ef797dcaa1b7804f5d1d2342cf5d2dc2ebe55f6828e47ee3335563ffd8d01fe
+size 370059
