+++ tools/render-assets/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "render-assets"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+toml = "0.5"
+serde = { version = "1.0", features = ["derive"] }
+indicatif = "0.17.1"
