+++ tools/render-assets/scripts/render-asset.py
